(threejs src) Cleaned up the escher cubemap code a bit. Also, I think it looks much nicer without illumination... ? At first I though the cube around the sphere was x-inverted (unless the in tention is to simulate refraction...) I've tried changing that but the effect is still weird. I guess those textures work mainly for the sphere and not for the panorama.
I reused the code and did another example using the skymap. More to come ;)

I can see some lines at the edges of the planes though. I guess that's antialiasing kicking in. Should be fixed with a proper cube mesh I guess.

CanvasRenderer has a property called `autoClear` which is true by default. By making it false you have control of when the renderer cleans the screen (by calling renderer.clear()). I wonder if something like that can be added to the WebGLRenderer? That way we can render the cube outside the spheres from a diferent camera (centered in the middle of the screen) than the spheres floating.

Also, added `shading` parameter to `MeshBasicMaterial` (environment mapping was being rendered faceted when using `MeshBasicMaterial` because `shading` was missing (and thus Faceted)).