(threejs src) Added shadows for DirectionalLights.
After shadow mapping fixes orthographic shadow projection finally started to work.

Work in progress. Still need to figure out some way to set up shadow camera automagically with respect to the actual camera (so that whole scene is shadowed when the camera moves). For the moment shadow camera needs to be handled manually (like for SpotLight).