(threejs src) Added new custom attribute binding type - "faceVertices".
This allows to give unique values for each face use of the vertex (as opposed to unique custom attributes per vertex or per face).

For the moment just Vector3 and Color attribute types are handled.

Primary motivation for this binding type is to be able to do barycentric wireframe rendering in shader.