(threejs src) Moving *Controls to examples/js
The reasoning is because I see a lot of people asking for tweaks so the control X works on their project.
The problem is that they can't just tweak the code because the controls is inside the lib.
By having them outside they should be able to guess they can customise the file for their project.

And yes, it also removes some bloat from the lib.