(threejs src) Added support for fog to Sprites.
Just set "sprite.fog = true" to get scene fog working on the sprite.

Internally, implementation is a bit different than for other objects: because there is just a single material for all Sprites, fog is always compiled into the shader and it's turned on/off (or fog type is changed) via uniform.