(threejs src) Editor: added handling of fog to "World" section.
Not sure if helpers should be fogged or not. It does look pretty weird when they are not fogged. Also having helpers fogged helps to get a sense of the fog effect over larger space.

Maybe there could be a checkbox for enabling this, to be able to see things if they need to be modified while in the fog.

Also found and fixed a bug deep in WebGLRenderer, where shaders were not recompiled when changing fog types because fog type wasn't taken into account when creating shader code hash. Phew, that was a tough one to figure out. This Editor thingie really forces us to dust all nooks and cranies of the codebase ;)