(threejs src) Added area weighted vertex normals computation to Geometry.
See discussion in #2613

Finally I just left it behind a flag. I was not able to get as good results for TextGeometry as with unweighted method and split materials for side and front faces.

I wonder how I got that nice look as in screenshot I posted in the issue. It still does look better with weighted than unweighted but nowhere so close as in that screenshot. Either I did some mistake when testing before or there is some bug in implementation now.